+++ src/Symfony/Component/Console/Descriptor/JsonDescriptor.php
@@ -40,6 +40,9 @@ class JsonDescriptor extends Descriptor
     protected function describeInputOption(InputOption $option, array $options = [])
     {
         $this->writeData($this->getInputOptionData($option), $options);
+        if ($option->isNegatable()) {
+            $this->writeData($this->getInputOptionData($option, true), $options);
+        }
     }
 
     /**
@@ -111,9 +114,17 @@ class JsonDescriptor extends Descriptor
         ];
     }
 
-    private function getInputOptionData(InputOption $option): array
+    private function getInputOptionData(InputOption $option, bool $negated = false): array
     {
-        return [
+        return $negated ? [
+            'name' => '--no-'.$option->getName(),
+            'shortcut' => '',
+            'accept_value' => false,
+            'is_value_required' => false,
+            'is_multiple' => false,
+            'description' => 'Negate the "--'.$option->getName().'" option',
+            'default' => false,
+        ] : [
             'name' => '--'.$option->getName(),
             'shortcut' => $option->getShortcut() ? '-'.str_replace('|', '|-', $option->getShortcut()) : '',
             'accept_value' => $option->acceptValue(),
@@ -134,6 +145,9 @@ class JsonDescriptor extends Descriptor
         $inputOptions = [];
         foreach ($definition->getOptions() as $name => $option) {
             $inputOptions[$name] = $this->getInputOptionData($option);
+            if ($option->isNegatable()) {
+                $inputOptions['no-'.$name] = $this->getInputOptionData($option, true);
+            }
         }
 
         return ['arguments' => $inputArguments, 'options' => $inputOptions];