+++ src/Symfony/Component/Console/Input/InputDefinition.php
@@ -33,6 +33,7 @@ class InputDefinition
     private $hasAnArrayArgument = false;
     private $hasOptional;
     private $options;
+    private $negations;
     private $shortcuts;
 
     /**
@@ -208,6 +209,7 @@ class InputDefinition
     {
         $this->options = [];
         $this->shortcuts = [];
+        $this->negations = [];
         $this->addOptions($options);
     }
 
@@ -245,6 +247,14 @@ class InputDefinition
             foreach (explode('|', $option->getShortcut()) as $shortcut) {
                 $this->shortcuts[$shortcut] = $option->getName();
             }
+        }
+
+        if ($option->isNegatable()) {
+            $negatedName = 'no-'.$option->getName();
+            if (isset($this->options[$negatedName])) {
+                throw new LogicException(sprintf('An option named "%s" already exists.', $negatedName));
+            }
+            $this->negations[$negatedName] = $option->getName();
         }
     }
 
@@ -298,6 +308,14 @@ class InputDefinition
     }
 
     /**
+     * Returns true if an InputOption object exists by negated name.
+     */
+    public function hasNegation(string $name): bool
+    {
+        return isset($this->negations[$name]);
+    }
+
+    /**
      * Gets an InputOption by shortcut.
      *
      * @return InputOption An InputOption object
@@ -339,6 +357,22 @@ class InputDefinition
     }
 
     /**
+     * Returns the InputOption name given a negation.
+     *
+     * @throws InvalidArgumentException When option given does not exist
+     *
+     * @internal
+     */
+    public function negationToName(string $negation): string
+    {
+        if (!isset($this->negations[$negation])) {
+            throw new InvalidArgumentException(sprintf('The "--%s" option does not exist.', $negation));
+        }
+
+        return $this->negations[$negation];
+    }
+
+    /**
      * Gets the synopsis.
      *
      * @return string The synopsis
@@ -362,7 +396,8 @@ class InputDefinition
                 }
 
                 $shortcut = $option->getShortcut() ? sprintf('-%s|', $option->getShortcut()) : '';
-                $elements[] = sprintf('[%s--%s%s]', $shortcut, $option->getName(), $value);
+                $negation = $option->isNegatable() ? sprintf('|--no-%s', $option->getName()) : '';
+                $elements[] = sprintf('[%s--%s%s%s]', $shortcut, $option->getName(), $value, $negation);
             }
         }